refactor: ajust some wrongs by type BigInteger to Integer

+++ back-poly/polyrate/src/main/java/nick/project/polyrate/Moeda/Repository/MoedaRepository.java
@@ -3,5 +3,5 @@ package nick.project.polyrate.Moeda.Repository;
 import nick.project.polyrate.Moeda.Entity.MoedaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MoedaRepository extends JpaRepository<MoedaEntity, Long> {
+public interface MoedaRepository extends JpaRepository<MoedaEntity, Integer> {
 }
